Support for new Timer API in Flink runner (streaming)
