+++ runners/flink/runner/src/main/java/org/apache/beam/runners/flink/translation/wrappers/streaming/WindowDoFnOperator.java
@@ -17,58 +17,30 @@
  */
 package org.apache.beam.runners.flink.translation.wrappers.streaming;
 
-
-import com.google.common.collect.HashMultiset;
-import com.google.common.collect.Multiset;
-import java.io.DataInputStream;
-import java.io.DataOutputStream;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.OutputStream;
 import java.nio.ByteBuffer;
 import java.util.Collection;
 import java.util.Collections;
-import java.util.Comparator;
-import java.util.HashMap;
-import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
-import java.util.PriorityQueue;
-import java.util.Queue;
-import java.util.Set;
-import java.util.concurrent.ScheduledFuture;
-import javax.annotation.Nullable;
-
-import org.apache.beam.runners.core.ExecutionContext;
 import org.apache.beam.runners.core.GroupAlsoByWindowViaWindowSetNewDoFn;
 import org.apache.beam.runners.core.KeyedWorkItem;
 import org.apache.beam.runners.core.KeyedWorkItems;
 import org.apache.beam.runners.core.SystemReduceFn;
-import org.apache.beam.runners.flink.translation.wrappers.DataInputViewWrapper;
 import org.apache.beam.sdk.coders.Coder;
 import org.apache.beam.sdk.options.PipelineOptions;
 import org.apache.beam.sdk.transforms.DoFn;
 import org.apache.beam.sdk.transforms.windowing.BoundedWindow;
-import org.apache.beam.sdk.util.TimeDomain;
 import org.apache.beam.sdk.util.TimerInternals;
 import org.apache.beam.sdk.util.WindowedValue;
 import org.apache.beam.sdk.util.WindowingStrategy;
 import org.apache.beam.sdk.util.state.StateInternals;
 import org.apache.beam.sdk.util.state.StateInternalsFactory;
-import org.apache.beam.sdk.util.state.StateNamespace;
 import org.apache.beam.sdk.util.state.TimerInternalsFactory;
 import org.apache.beam.sdk.values.KV;
 import org.apache.beam.sdk.values.PCollectionView;
 import org.apache.beam.sdk.values.TupleTag;
 import org.apache.flink.api.common.typeinfo.TypeInformation;
 import org.apache.flink.api.java.tuple.Tuple2;
-import org.apache.flink.core.memory.DataInputView;
-import org.apache.flink.runtime.state.AbstractStateBackend;
-import org.apache.flink.runtime.state.StateHandle;
-import org.apache.flink.streaming.api.watermark.Watermark;
-import org.apache.flink.streaming.runtime.operators.Triggerable;
-import org.apache.flink.streaming.runtime.tasks.StreamTaskState;
-import org.joda.time.Instant;
 
 /**
  * Flink operator for executing window {@link DoFn DoFns}.
@@ -77,20 +49,8 @@ import org.joda.time.Instant;
  * @param <OutputT>
  */
 public class WindowDoFnOperator<K, InputT, OutputT>
-    extends DoFnOperator<KeyedWorkItem<K, InputT>, KV<K, OutputT>, WindowedValue<KV<K, OutputT>>>
-    implements Triggerable {
-
-  private final TimerInternals.TimerDataCoder timerCoder;
-
-  private transient Set<Tuple2<ByteBuffer, TimerInternals.TimerData>> watermarkTimers;
-  private transient Queue<Tuple2<ByteBuffer, TimerInternals.TimerData>> watermarkTimersQueue;
-
-  private transient Queue<Tuple2<ByteBuffer, TimerInternals.TimerData>> processingTimeTimersQueue;
-  private transient Set<Tuple2<ByteBuffer, TimerInternals.TimerData>> processingTimeTimers;
-  private transient Multiset<Long> processingTimeTimerTimestamps;
-  private transient Map<Long, ScheduledFuture<?>> processingTimeTimerFutures;
-
-  private transient FlinkTimerInternals timerInternals;
+    extends DoFnOperator<KeyedWorkItem<K, InputT>, KV<K, OutputT>,
+    WindowedValue<KV<K, OutputT>>> {
 
   private final SystemReduceFn<K, InputT, ?, OutputT, BoundedWindow> systemReduceFn;
 
@@ -119,8 +79,6 @@ public class WindowDoFnOperator<K, InputT, OutputT>
 
     this.systemReduceFn = systemReduceFn;
 
-    this.timerCoder =
-        TimerInternals.TimerDataCoder.of(windowingStrategy.getWindowFn().windowCoder());
   }
 
   @Override
@@ -152,383 +110,12 @@ public class WindowDoFnOperator<K, InputT, OutputT>
     return doFn;
   }
 
-
   @Override
-  public void open() throws Exception {
-
-    // might already be initialized from restoreTimers()
-    if (watermarkTimers == null) {
-      watermarkTimers = new HashSet<>();
-
-      watermarkTimersQueue = new PriorityQueue<>(
-          10,
-          new Comparator<Tuple2<ByteBuffer, TimerInternals.TimerData>>() {
-            @Override
-            public int compare(
-                Tuple2<ByteBuffer, TimerInternals.TimerData> o1,
-                Tuple2<ByteBuffer, TimerInternals.TimerData> o2) {
-              return o1.f1.compareTo(o2.f1);
-            }
-          });
-    }
-
-    if (processingTimeTimers == null) {
-      processingTimeTimers = new HashSet<>();
-      processingTimeTimerTimestamps = HashMultiset.create();
-      processingTimeTimersQueue = new PriorityQueue<>(
-          10,
-          new Comparator<Tuple2<ByteBuffer, TimerInternals.TimerData>>() {
-            @Override
-            public int compare(
-                Tuple2<ByteBuffer, TimerInternals.TimerData> o1,
-                Tuple2<ByteBuffer, TimerInternals.TimerData> o2) {
-              return o1.f1.compareTo(o2.f1);
-            }
-          });
-    }
-
-    // ScheduledFutures are not checkpointed
-    processingTimeTimerFutures = new HashMap<>();
-
-    timerInternals = new FlinkTimerInternals();
-
-    // call super at the end because this will call getDoFn() which requires stateInternals
-    // to be set
-    super.open();
-  }
-
-  @Override
-  protected ExecutionContext.StepContext createStepContext() {
-    return new WindowDoFnOperator.StepContext();
-  }
-
-  private void registerEventTimeTimer(TimerInternals.TimerData timer) {
-    Tuple2<ByteBuffer, TimerInternals.TimerData> keyedTimer =
-        new Tuple2<>((ByteBuffer) getStateBackend().getCurrentKey(), timer);
-    if (watermarkTimers.add(keyedTimer)) {
-      watermarkTimersQueue.add(keyedTimer);
-    }
-  }
-
-  private void deleteEventTimeTimer(TimerInternals.TimerData timer) {
-    Tuple2<ByteBuffer, TimerInternals.TimerData> keyedTimer =
-        new Tuple2<>((ByteBuffer) getStateBackend().getCurrentKey(), timer);
-    if (watermarkTimers.remove(keyedTimer)) {
-      watermarkTimersQueue.remove(keyedTimer);
-    }
-  }
-
-  private void registerProcessingTimeTimer(TimerInternals.TimerData timer) {
-    Tuple2<ByteBuffer, TimerInternals.TimerData> keyedTimer =
-        new Tuple2<>((ByteBuffer) getStateBackend().getCurrentKey(), timer);
-    if (processingTimeTimers.add(keyedTimer)) {
-      processingTimeTimersQueue.add(keyedTimer);
-
-      // If this is the first timer added for this timestamp register a timer Task
-      if (processingTimeTimerTimestamps.add(timer.getTimestamp().getMillis(), 1) == 0) {
-        ScheduledFuture<?> scheduledFuture = registerTimer(timer.getTimestamp().getMillis(), this);
-        processingTimeTimerFutures.put(timer.getTimestamp().getMillis(), scheduledFuture);
-      }
-    }
-  }
-
-  private void deleteProcessingTimeTimer(TimerInternals.TimerData timer) {
-    Tuple2<ByteBuffer, TimerInternals.TimerData> keyedTimer =
-        new Tuple2<>((ByteBuffer) getStateBackend().getCurrentKey(), timer);
-    if (processingTimeTimers.remove(keyedTimer)) {
-      processingTimeTimersQueue.remove(keyedTimer);
-
-      // If there are no timers left for this timestamp, remove it from queue and cancel the
-      // timer Task
-      if (processingTimeTimerTimestamps.remove(timer.getTimestamp().getMillis(), 1) == 1) {
-        ScheduledFuture<?> triggerTaskFuture =
-            processingTimeTimerFutures.remove(timer.getTimestamp().getMillis());
-        if (triggerTaskFuture != null && !triggerTaskFuture.isDone()) {
-          triggerTaskFuture.cancel(false);
-        }
-      }
-
-    }
-  }
-
-  @Override
-  public void trigger(long time) throws Exception {
-
-    //Remove information about the triggering task
-    processingTimeTimerFutures.remove(time);
-    processingTimeTimerTimestamps.setCount(time, 0);
-
-    boolean fire;
-
-    do {
-      Tuple2<ByteBuffer, TimerInternals.TimerData> timer = processingTimeTimersQueue.peek();
-      if (timer != null && timer.f1.getTimestamp().getMillis() <= time) {
-        fire = true;
-
-        processingTimeTimersQueue.remove();
-        processingTimeTimers.remove(timer);
-
-        setKeyContext(timer.f0);
-
-        pushbackDoFnRunner.processElement(WindowedValue.valueInGlobalWindow(
-            KeyedWorkItems.<K, InputT>timersWorkItem(
-                (K) stateInternals.getKey(),
-                Collections.singletonList(timer.f1))));
-
-      } else {
-        fire = false;
-      }
-    } while (fire);
-
-  }
-
-  @Override
-  public void processWatermark(Watermark mark) throws Exception {
-    processWatermark1(mark);
-  }
-
-  @Override
-  public void processWatermark1(Watermark mark) throws Exception {
-    pushbackDoFnRunner.startBundle();
-
-    this.currentInputWatermark = mark.getTimestamp();
-
-    // hold back by the pushed back values waiting for side inputs
-    long actualInputWatermark = Math.min(getPushbackWatermarkHold(), mark.getTimestamp());
-
-    boolean fire;
-
-    do {
-      Tuple2<ByteBuffer, TimerInternals.TimerData> timer = watermarkTimersQueue.peek();
-      if (timer != null && timer.f1.getTimestamp().getMillis() < actualInputWatermark) {
-        fire = true;
-
-        watermarkTimersQueue.remove();
-        watermarkTimers.remove(timer);
-
-        setKeyContext(timer.f0);
-
-        pushbackDoFnRunner.processElement(WindowedValue.valueInGlobalWindow(
-                KeyedWorkItems.<K, InputT>timersWorkItem(
-                    (K) stateInternals.getKey(),
-                    Collections.singletonList(timer.f1))));
-
-      } else {
-        fire = false;
-      }
-    } while (fire);
-
-    Instant watermarkHold = stateInternals.watermarkHold();
-
-    long combinedWatermarkHold = Math.min(watermarkHold.getMillis(), getPushbackWatermarkHold());
-
-    long potentialOutputWatermark = Math.min(currentInputWatermark, combinedWatermarkHold);
-
-    if (potentialOutputWatermark > currentOutputWatermark) {
-      currentOutputWatermark = potentialOutputWatermark;
-      output.emitWatermark(new Watermark(currentOutputWatermark));
-    }
-    pushbackDoFnRunner.finishBundle();
-
-  }
-
-  @Override
-  public StreamTaskState snapshotOperatorState(long checkpointId, long timestamp) throws Exception {
-    StreamTaskState result = super.snapshotOperatorState(checkpointId, timestamp);
-
-    AbstractStateBackend.CheckpointStateOutputView outputView =
-        getStateBackend().createCheckpointStateOutputView(checkpointId, timestamp);
-
-    snapshotTimers(outputView);
-
-    StateHandle<DataInputView> handle = outputView.closeAndGetHandle();
-
-    // this might overwrite stuff that super checkpointed
-    result.setOperatorState(handle);
-
-    return result;
-  }
-
-  @Override
-  public void restoreState(StreamTaskState state) throws Exception {
-    super.restoreState(state);
-
-    @SuppressWarnings("unchecked")
-    StateHandle<DataInputView> operatorState =
-        (StateHandle<DataInputView>) state.getOperatorState();
-
-    DataInputView in = operatorState.getState(getUserCodeClassloader());
-
-    restoreTimers(new DataInputViewWrapper(in));
-  }
-
-  private void restoreTimers(InputStream in) throws IOException {
-    DataInputStream dataIn = new DataInputStream(in);
-
-    int numWatermarkTimers = dataIn.readInt();
-
-    watermarkTimers = new HashSet<>(numWatermarkTimers);
-
-    watermarkTimersQueue = new PriorityQueue<>(
-        Math.max(numWatermarkTimers, 1),
-        new Comparator<Tuple2<ByteBuffer, TimerInternals.TimerData>>() {
-          @Override
-          public int compare(
-                  Tuple2<ByteBuffer, TimerInternals.TimerData> o1,
-                  Tuple2<ByteBuffer, TimerInternals.TimerData> o2) {
-            return o1.f1.compareTo(o2.f1);
-          }
-        });
-
-    for (int i = 0; i < numWatermarkTimers; i++) {
-      int length = dataIn.readInt();
-      byte[] keyBytes = new byte[length];
-      dataIn.readFully(keyBytes);
-      TimerInternals.TimerData timerData = timerCoder.decode(dataIn, Coder.Context.NESTED);
-      Tuple2<ByteBuffer, TimerInternals.TimerData> keyedTimer =
-          new Tuple2<>(ByteBuffer.wrap(keyBytes), timerData);
-      if (watermarkTimers.add(keyedTimer)) {
-        watermarkTimersQueue.add(keyedTimer);
-      }
-    }
-
-    int numProcessingTimeTimers = dataIn.readInt();
-
-    processingTimeTimers = new HashSet<>(numProcessingTimeTimers);
-    processingTimeTimersQueue = new PriorityQueue<>(
-        Math.max(numProcessingTimeTimers, 1),
-        new Comparator<Tuple2<ByteBuffer, TimerInternals.TimerData>>() {
-          @Override
-          public int compare(
-              Tuple2<ByteBuffer, TimerInternals.TimerData> o1,
-              Tuple2<ByteBuffer, TimerInternals.TimerData> o2) {
-            return o1.f1.compareTo(o2.f1);
-          }
-        });
-
-    processingTimeTimerTimestamps = HashMultiset.create();
-    processingTimeTimerFutures = new HashMap<>();
-
-    for (int i = 0; i < numProcessingTimeTimers; i++) {
-      int length = dataIn.readInt();
-      byte[] keyBytes = new byte[length];
-      dataIn.readFully(keyBytes);
-      TimerInternals.TimerData timerData = timerCoder.decode(dataIn, Coder.Context.NESTED);
-      Tuple2<ByteBuffer, TimerInternals.TimerData> keyedTimer =
-          new Tuple2<>(ByteBuffer.wrap(keyBytes), timerData);
-      if (processingTimeTimers.add(keyedTimer)) {
-        processingTimeTimersQueue.add(keyedTimer);
-
-        //If this is the first timer added for this timestamp register a timer Task
-        if (processingTimeTimerTimestamps.add(timerData.getTimestamp().getMillis(), 1) == 0) {
-          // this registers a timer with the Flink processing-time service
-          ScheduledFuture<?> scheduledFuture =
-              registerTimer(timerData.getTimestamp().getMillis(), this);
-          processingTimeTimerFutures.put(timerData.getTimestamp().getMillis(), scheduledFuture);
-        }
-
-      }
-    }
-  }
-
-  private void snapshotTimers(OutputStream out) throws IOException {
-    DataOutputStream dataOut = new DataOutputStream(out);
-    dataOut.writeInt(watermarkTimersQueue.size());
-    for (Tuple2<ByteBuffer, TimerInternals.TimerData> timer : watermarkTimersQueue) {
-      dataOut.writeInt(timer.f0.limit());
-      dataOut.write(timer.f0.array(), 0, timer.f0.limit());
-      timerCoder.encode(timer.f1, dataOut, Coder.Context.NESTED);
-    }
-
-    dataOut.writeInt(processingTimeTimersQueue.size());
-    for (Tuple2<ByteBuffer, TimerInternals.TimerData> timer : processingTimeTimersQueue) {
-      dataOut.writeInt(timer.f0.limit());
-      dataOut.write(timer.f0.array(), 0, timer.f0.limit());
-      timerCoder.encode(timer.f1, dataOut, Coder.Context.NESTED);
-    }
-  }
-
-  /**
-   * {@link StepContext} for running {@link DoFn DoFns} on Flink. This does now allow
-   * accessing state or timer internals.
-   */
-  protected class StepContext extends DoFnOperator.StepContext {
-
-    @Override
-    public TimerInternals timerInternals() {
-      return timerInternals;
-    }
-  }
-
-  private class FlinkTimerInternals implements TimerInternals {
-
-    @Override
-    public void setTimer(
-            StateNamespace namespace, String timerId, Instant target, TimeDomain timeDomain) {
-      throw new UnsupportedOperationException("Setting a timer by ID is not yet supported.");
-    }
-
-    @Deprecated
-    @Override
-    public void setTimer(TimerData timerKey) {
-      if (timerKey.getDomain().equals(TimeDomain.EVENT_TIME)) {
-        registerEventTimeTimer(timerKey);
-      } else if (timerKey.getDomain().equals(TimeDomain.PROCESSING_TIME)) {
-        registerProcessingTimeTimer(timerKey);
-      } else {
-        throw new UnsupportedOperationException(
-                "Unsupported time domain: " + timerKey.getDomain());
-      }
-    }
-
-    @Deprecated
-    @Override
-    public void deleteTimer(StateNamespace namespace, String timerId) {
-      throw new UnsupportedOperationException(
-              "Canceling of a timer by ID is not yet supported.");
-    }
-
-    @Override
-    public void deleteTimer(StateNamespace namespace, String timerId, TimeDomain timeDomain) {
-      throw new UnsupportedOperationException(
-          "Canceling of a timer by ID is not yet supported.");
-    }
-
-    @Deprecated
-    @Override
-    public void deleteTimer(TimerData timerKey) {
-      if (timerKey.getDomain().equals(TimeDomain.EVENT_TIME)) {
-        deleteEventTimeTimer(timerKey);
-      } else if (timerKey.getDomain().equals(TimeDomain.PROCESSING_TIME)) {
-        deleteProcessingTimeTimer(timerKey);
-      } else {
-        throw new UnsupportedOperationException(
-                "Unsupported time domain: " + timerKey.getDomain());
-      }
-    }
-
-    @Override
-    public Instant currentProcessingTime() {
-      return new Instant(getCurrentProcessingTime());
-    }
-
-    @Nullable
-    @Override
-    public Instant currentSynchronizedProcessingTime() {
-      return new Instant(getCurrentProcessingTime());
-    }
-
-    @Override
-    public Instant currentInputWatermarkTime() {
-      return new Instant(Math.min(currentInputWatermark, getPushbackWatermarkHold()));
-    }
-
-    @Nullable
-    @Override
-    public Instant currentOutputWatermarkTime() {
-      return new Instant(currentOutputWatermark);
-    }
-
+  public void fireTimer(Tuple2<ByteBuffer, TimerInternals.TimerData> timer) {
+    pushbackDoFnRunner.processElement(WindowedValue.valueInGlobalWindow(
+        KeyedWorkItems.<K, InputT>timersWorkItem(
+            (K) stateInternals.getKey(),
+            Collections.singletonList(timer.f1))));
   }
 
 }
